Changes to scripting
Raycasting still not working as I want it too.
The problem is mainly with getting the position and rotation to work so
that the raycast doesn't clip into the users body. Other than that my
testing shows that it should work smoothly.

Movement works repetitively but it doesn't queue tasks to be done one
after the other. So it will navigate to an item but won't then navigate
to the player. This is a problem with the scheduling of tasks so I think
I need to do an overhaul of how it's done later on but for now just
putting seperate methods to chase an object and then after that method
finishes then chase the player would work for our purposes.

Also minor bug is that movement continues to work after death.