AI UPGRADED
AI now will run towards the closest weapon, will try to pick it up. Main
caveats are that the enemy doesn't get quite close enough to pick up the
weapon, which can be fixed by buffing the range that the weapon gets
picked up from as well as the pathfinding is a bit buggy now so i need
to go back and check it.