Backpack slots and Hover Text
Made the backpack slot work

Created hovertext:

To give a ui element a hovertext, give it a EVENT TRIGGER and a HOVER
TEXT component. Make the event trigger call the hover text's display
function on a OnPointerEnter event. In the hover text component, set
which AUTOTEXT component you want to display the text, and set the text
itself.