Screw it all!!!!!!
So...the pathfinding works but it makes the units go wacky and off to
the side of the screen. If you increase A*'s grid size you can see where
the AI is actually pathfinding too.

I've got work tomorrow but basically I believe that it's pathfinding
that needs to be fixed.

If Deon could look at it and do what he did before with the SimplyPath,
to see if that's still the problem, I'd be very grateful, otherwise
tomorrow night I'll start on just hard programming in all the features.