Ai system work
All items now have a threat value (0-1)
all contestants have confidence (-1 - 1) which slowly builds up over
time (currently no way to lose it)
When contestants are randomly moving, they will tend to the inside of
the arena if they are confident, outside if they are not.
use the third skill to see a contestant's status, confidence and threat.