Character Model Implementation Complete
All previous bugs with the implementation of the new model are now
fixed.

All enemies and guards now share the new model. A current bug with
enemies and guards is that they do not change their colour depending on
the enemy type.

Ragdoll physics are in place but currently disabled, a future script
will need to handle on death activate the model's skeleton and disabling
animator for ragdoll physics to work efficiently.